+++ experimental/renderer/galaxy_big/BUILD
@@ -18,11 +18,11 @@ rust_binary(
     deps = [
         ":galaxy_big_renderer",
         "//core:init",
-        "//core/geom:geometry",
+        "//core/geom",
         "//core/geom/gen:icosphere",
         "//core/rend/vk:sdl2_vulkan_interop",
+        "//core/rend/vk:vk_base_renderer",
         "//core/rend/vk:vk_lite",
-        "//core/rend/vk:vk_renderer",
         "//experimental/physics:fast_cosmic_physics",
         "//third_party/cargo:cgmath",
         "//third_party/cargo:chrono",
@@ -48,8 +48,9 @@ rust_library(
         "-C opt-level=2",
     ],
     deps = [
-        "//core/geom:geometry",
+        "//core/geom",
         "//core/geom/gen:icosphere",
+        "//core/rend/vk:vk_base_renderer",
         "//core/rend/vk:vk_buffer_cache",
         "//core/rend/vk:vk_buffer_support",
         "//core/rend/vk:vk_descriptor_support",
@@ -57,7 +58,6 @@ rust_library(
         "//core/rend/vk:vk_instance_support",
         "//core/rend/vk:vk_lite",
         "//core/rend/vk:vk_pipeline_support",
-        "//core/rend/vk:vk_renderer",
         "//core/rend/vk:vk_swapchain_support",
         "//third_party/cargo:cgmath",
         "//third_party/cargo:log",
