+++ experimental/renderer/planet/BUILD
@@ -20,8 +20,11 @@ rust_library(
         "-C opt-level=2",
     ],
     deps = [
-        "//core/geom:geometry",
+        "//core/geom",
+        "//core/geom:geom_vk_support",
         "//core/geom/gen:icosphere",
+        "//core/rend/vk:vk_application",
+        "//core/rend/vk:vk_base_renderer",
         "//core/rend/vk:vk_buffer_cache",
         "//core/rend/vk:vk_buffer_support",
         "//core/rend/vk:vk_descriptor_support",
@@ -29,7 +32,6 @@ rust_library(
         "//core/rend/vk:vk_instance_support",
         "//core/rend/vk:vk_lite",
         "//core/rend/vk:vk_pipeline_support",
-        "//core/rend/vk:vk_renderer",
         "//core/rend/vk:vk_swapchain_support",
         "//third_party/cargo:cgmath",
         "//third_party/cargo:log",
@@ -46,7 +48,7 @@ rust_library(
     ],
     deps = [
         ":planet_renderer",
-        "//core/geom:geometry",
+        "//core/geom",
         "//core/geom/gen:icosphere",
         "//third_party/cargo:chrono",
         "//third_party/cargo:log",
